feat: enhance contract event handling and refactor indexer services
- Implemented a retry mechanism in GetContractEvents to handle timeouts and ensure successful retrieval of contract events.
- Updated WaitForTransactionMined to include error handling for transaction failures.
- Removed unused indexer test and indexer service files to streamline the codebase.
- Added new migration for engine bump to support upcoming features.
